requirements.txt
python-telegram-bot==20.7
google-generativeai>=0.8.5
python-dotenv
